ci: Clean up name service state
The partition test input is a subdirectory of a normal test input. A
possible but unconfirmed race would be if these tests ran concurrently.
In this case, it is possible for the test setup fixture for the normal
test to delete the input of the partition test. Instead, change the
partition test input to be a sibling of a normal test input.

Also, etcd state is preserved across test invocations. As a precaution,
reset etcd state before testing to guard against any erroneous etcd
state from leaking between test instances.

[REWRITTEN COMMIT: self-contained subtree]