Fix race between tsuba::Create and tsuba::Open
In a recent commit which fixed a race between RDG creation and checking,
I turned RegisterIfAbsent from an opportunistic activity to something
that would return an error if it failed. This turned a race between
tsuba::Create and tsuba::Open that was benign and ignored into a visible
API error.

Briefly, the race is this:

  PropertyFileGraph::Write():
    tsuba::Create():
      if (ID == 0) {
        tsuba::FileStore(....meta file....)  // (A)
      }
      CreateIfAbsent()

    tsuba::Open():
      RegisterIfAbsent():
        FindLatestMetaFile():
          if (no meta file) {  // (B)
            return InvalidArgument
          }
        CreateIfAbsent()

Host 0 writes a meta file (A) but it is possible for host 1 to proceed
to FindLatestMetaFile before the meta file is created (B). This issue
doesn't happen with HttpNameServerClient because CreateIfAbsent in that
situation invokes a barrier between hosts.

The solution to this is to make name server operations always collective
operations.

While looking at this code, I also noticed that RegisterIfAbsent is
called recursively through RDGMeta::Make:

  RDGMeta::Make():
    RegisterIfAbsent():
      RDGMeta::Make():
        ...

RegisterIfAbsent is a collective operation but RDGMeta::Make looks like
an innocuous constructor. In order to reduce the chance of nested
collective operations, I've pulled up RegisterIfAbsent to the main tsuba
API entry points.

[REWRITTEN COMMIT: self-contained subtree]