O(1) random initialization and fallback
An intermittent issue is failure to initialize our random number
generators:

  random_device: rdseed failed

Because we try to be "more-correct" when initializing our generators, we
consume about 600 32-bits of entropy per thread. Depending on the STL
implementation, the underlying library may or may not be able to provide
this amount of entropy at once, consistently.

Instead, maintain a single pool of entropy for the system and initialize
threads from this pool. If we cannot initialize the system pool fully,
fall back to partial state initialization using the system clock.

Since we are not concerned with cryptographically secure applications,
this hopefully strikes the right balance between randomness and
crashing.