Generate Ubuntu packages
As part of a larger effort to simplify packaging, reduce our runtime
dependencies to only libraries that are available from the standard
sources of a distribution.

The main sources of packaging issues are:

1. libarrow. It is not widely available and it has major API version
  changes every few months.
2. libcurl+openssl. The specific libcrypto used varies between Ubuntu
  16.04 through 20.04.
3. MPI. MPI is source compatible within a major version but generally MPI
  programs linked against one MPI implementation cannot run with a
  different implementation.
4. libstdc++. For distributions that do support gcc-7 (the first version
  of gcc that supports C++17), we need to provide a compatible C++
  runtime. libstdc++ in turn depends on libc, so there is a limit to how
  old an distribution we can support.

Address issues (1) and (2) with static linkage.

Address issue (3) by distributing and linking against our own copy of
libopenmpi. This also sidesteps the quite old versions of libopenmpi
that are common on Ubuntu 16.04 (libopenmpi 1.10), Ubuntu 18.04
(libopenmpi 2) and Ubuntu 20.04 (libopenmpi 3). libopenmpi 1.10 and
libopenmpi 2 notably are not usually compiled to support
MPI_THREAD_MULTIPLE.

Distribute and link with libopenmpi4.

Various vendors do provide more up to date versions of libopenmpi. For
example, the AWS Ubuntu 18.04 distribution comes with a custom
libopenmpi4-aws variant. In the future, it might make sense to create a
separate libopenmpi4-katana package that fills in only if a user doesn't
have an alternative. Though given the lack of binary portability of MPI
programs, there might be a limit to how tolerant we can be.

Issue (4) affects Ubuntu 16.04 packaging which is on the outside of what
we intend to support. In this case, just copy over libstdc++ from a
backported g++-7 toolchain. Hopefully, we do not need to support this
hack for long.

[REWRITTEN COMMIT: self-contained subtree]