Update arrow dependency
Arrow v2 seems to be source compatible with the subset of Arrow 0.17/1.0
we have been using so far. Simplify internal development by statically
linking libarrow by default. We still need to support shared libraries
for conda, as among the binary distribution platforms we support:
ubuntu, centos and conda, it is the only one that has mainline support
for libarrow.

[REWRITTEN COMMIT: self-contained subtree]