Update conda version requirements
Galois works with libarrow1.0 and libarrow2.0. These two versions are
not compatible with each other, so we need to create a separate variant
of the galois package for each version of libarrow we want to support.

libarrow2.0 requires fmt>=7. Along the way, loosen the requirements for
header-only libraries like fmt, nlohmann_json and howardhinnant_date.

While adjusting version requirements for arrow, I encountered a somewhat
cryptic failure where the galois package was built but it was not
compatible with the galois-python specification, which led conda to
install a previous version of galois from the katanagraph channel. This
led to a compilation error as symbols had changed between the previous
version of libgalois and the current one.

There is probably a better way to organize conda subpackages to avoid
this subtle failure mode, but I didn't attempt it here.

[REWRITTEN COMMIT: self-contained subtree]