tsuba: move internal data structures out of public includes
~One functional change: Removed support for `parquet` partition files.
I don't think we have any graphs yet with that style of metadata.~

I decided to opt for the, less disruptive, try JSON first aproach.

The rest should all be refactoring. I moved the include files for data
structures that the rest of the codebase has no business looking at into
`src` and cleaned up the interfaces so that the only way to construct
the remaining structures is via `RDGHandle`s. This forces everyone to
call open first; which will be important when calling open grabs a
lease.

Hopefully this makes it clear what the actual public interfaces to
`tsuba` are.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>

[REWRITTEN COMMIT: self-contained subtree]