Rename galois to katana (1/2)
This is a two part change in order to not confuse git with simultaneous
file renaming and update.

This commit is the result of applying:

  scripts/rename_galois_to_katana.sh files

The overall project is now called katana, and the C++ namespace has been
updated to match. Galois is now simply the name of one component of the
katana project. libgalois is the data structure and scheduling support
library to help users write safe parallel programs.

I also flattened several nested C++ namespaces (galois::worklists,
galois::runtime, galois::graphs, galois::substrate) into their parent.
Having many nested namespaces is an anti-pattern in C++. If developers
want to organize their code, they can use directories or separate
library components.

In light of the focus of libgalois as a data structure and scheduling
support library, I removed the subdirectories galois/worklists,
galois/runtime, galois/graphs galois/substrate as these distinctions are
not particularly relevant to users of libgalois.