graph-convert: support out-of-core GR->KG conversions
I do this by reading the GR header the performing a partial copy where
all of the GR file except for the node data are copied into an RDG
directory. The result is a valid topology file which is then connected
to the RDG Metadata.

This is exposed in graph convert as a special case optimization for the
case where the edge type is void.

```
$ graph-convert --edgeType=void -gr2kg \
    s3://example-graphs/road-europe.gr s3://new_graph/rdg
```

Caveats:
 * doesn't work across different back-ends (e.g. both source and dest
   must be in the local file system or in s3, gs etc).
 * When in s3 both source and dest have to be in the same region
 * Azure implementation is still TODO

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>