Avoid weak vtables
If a virtual class has no out-of-line definitions, its vtable is emitted
in each translation unit it is referenced. RTTI tables follow the same
pattern so if we want consistent type info across users of a library
like libgalois, we should avoid having multiple vtables for the same
class.

Only clang has a diagnostic for this issue.

I also considered applying this throughout the codebase but it turns out
some external dependencies do not compile with -Wweak-vtable at all.