Rename NodeProperty/NodeTable to node_property
Properties are tables but they are really a specific kind of table: one
where rows correspond to nodes or edges of a property graph. Match the
API with this mental model.

I debated between PropertyFileGraph::GetNodeProperty(int i) and
PropertyFileGraph::node_property(int i).

On balance, I favored GetNodeProperty(int i) because there is a
corresponding GetNodePropertyTyped method that is certainly not a field
access and thus isn't a good candidate to be named node_property_typed,
and it seemed reasonable to me to keep these two related accessors with
the same naming style.