Use C++11 attribute syntax for KATANA_EXPORT
Both gcc and mvcc on Windows support mixing __attribute__((foo)) and
[[foo]] syntax, but on Linux, at least gcc-7 cannot. Both gcc and clang
do support [[gnu::visibility]] as an alternative. Switch to using this
to allow declarations with multiple attributes.

An issue when switching to standard attribute syntax is that the C++
standard is more restrictive about attribute placement. Attributes must
appear before return values in function declarations and after
class/struct in type definitions.

A subtle case is attributes with friend. Attributes may only appear with
friend on a function definition and not a declaration.