Impove error output
Fix two bugs:

One, the error context is not reset upon creation of new errors, which
leads to an accumulation of unrelated error messages. Instead, reset the
state of the error info thread-local buffer when a new error info is
created.

Two, when an error is created without a context and then a context is
added, e.g.,

  ErrorInfo ei = ErrorCode::InvalidArgument;
  ei = ei.WithContext("more");

the resulting message is simply "more". Instead prepend some
representation of the original error like "more: invalid argument".