Make libsupport shared always
While debugging an assertion failure in Result.cpp, I noticed that is
possible for a program to have multiple definitions of libsupport. For
example,

  my-lib.so -> libsupport.a
  application -> ... -> libsupport.a
  application -> my-lib.so

While previously this was benign though duplicative; now it can create
more serious issues because katana::Result uses a thread-local buffer to
maintain additional error information. If this buffer can be defined
multiple times, error context information can be accumulated in
different places depending on which version of libsupport a caller
referenced.

Given that, at the end of the day, the purpose of libsupport is to
provide a place for common functionality, it seems simpler to just make
libsupport always a shared library, so there will only be one version of
it for any application.