scripts/version: Handle different remotes in open and enterprise repos.
This can occur, for example, when a developer only makes changes in
the enterprise repo, and configures their open submodule to have a
only an origin remote and no personal fork while still using a
personal fork of the enterprise repo. This commit changes the
script to allow different setups of this kind with different
remote configurations for open and enterprise.