Use cmake CLI to build and install in conda build.
For reasons I do not understand and cannot reproduce, make install
was causing a complete recompile in he conda build. My only guess
is that some option in the conda environment causes a -D to change
in the build flags during installation forcing a recompile. But,
honestly that seems super unlikely.