Introduce Types to PropertyGraph
Differentiate between types and non-type/arbitrary properties.
Currently, assumes all boolean or uint8 properties are types.

Constructs a new array to store types of all nodes and another
for all edges. The types of a node are encoded using a TypeID.

Encode TypeID to identify/contain a unique combinations of
types. Currently, the encoding uses 8 bits, so it supports up
to 254 unique combinations of types (excluding unknown and
invalid types).