Implement bfs using direction optimization
This commit implements and adds direction optimization to Lonestar bfs.
This bfs version should be updated for stable/clear graph transposing.
Direction optimization requires both in- and out- edges.
Due to the lack of support on bi-directional iterations on PropertyGraph,
it creates new additional transposed graph on application level.
To avoid memory leak, it creates LargeArray at application level, which
could be problematic.