Bitset for edges for BC-level
Adds a bitset to mark active edges for bc-level. What this does is make
it so the backward Brandes step does not have to read the edge
destination as well as the distance label on the destination to
determine if the edge is part of the successors list of that node.