Use two labels of parent and distance on bfs-async (#315)
The past bfs-async only computes depths of bfs-tree from
a start node. This is different from bfs-direction-optimization
which computes parents of a node, differentiates semantics
of bfs variants, and diverges codes including verification,
and validation. This commit makes bfs-async compute
not only depths, but also parents, and therefore, integrates
semantics of bfs.

Co-authored-by: Hochan Lee <hochan@katanagraph.com>