RDG views support with RDGManifest (#271)
Transitions from RDGMeta to RDGManifest. RDGManifests encode the RDG view type in the name of the metadata (partition files and manifest files). This enables support for the notion of graph views, allowing for the persistence of multiple graphs alongside each other.

A tool is provided (scripts/migrate_rdgs.py) to help migrate RDGs to the new format. Invoke it with: './scripts/migrate_rdgs.py --fs /path/to/rdg --s3 s3://bucket-with-rdg/dirs'.

The NameServer is also deprecated with this PR.