galois/PropertyGraph: tweak interfaces to allow updates
I want to support updates of properties that do not change the topology
without repartitioning. To do that, I needed to tweak a few things.

The first is that I need to be able to un-set a property. To support
that I added a `SetNull` member to `ArrowRandomAccessBuilder`.

The second is that when I tweak a property graph then try to build an
attributed graph from it, this type labels function is called again. I
modified the semantics to re-compute the types rather than complain
since I believe those semantics are correct. When I update the
properties I may have changed some labels so types should be
re-computed.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>