Parquet{Read,Writ}er: block files when element count would be too large
This is a work around for an apparent limitation in the arrow library
when translating from parquet

String arrays have a de-facto limit in the number of elements they can
contain (it seems close to the number of string bytes). You can end up
with more elements than string bytes if there are mostly null entries.

We can account for this in memory by chunking arrays, but parquet does
not have this limitation and the chunking is lost in translation. Then
when the arrays are translated back into arrow Sting Builder complains
that there are too many elements for the array.

The workaround here chunks parquet files as well.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>