Add SIGPIPE handler
I debugged some SIGPIPE errors that eventually led to some configuration
options in libcurl to ignore SIGPIPE. While it is possible to enable
such options in each dependency we use if they are available, signal
handling is not reentrant or encapsulated, and it is traditionally the
responsibility of application program and not libraries to set them up.

Instead, refactor backtrace support to be the place for general signal
handling and add a "log and ignore" handler for SIGPIPE, which is a
common option for long-running programs.