[KAT-972] Visit arrow multiple args (#387)
Adds support for the VisitArrow function to take multiple arguments (note: this ability requires switching the order of the arguments, though the original order is maintained for backwards compatibility). This allows one to call VisitArrow(Visitor visitor, ArrowType1 at1, ArrowType2 at2, ...), which will then in turn call the correct specialization of visitor.Call(...).