tsuba: keep all type information around
As others have noted (@witchel) it is desirable to have type information
for properties that are not loaded. This PR adds that facility along
with some mechanism for the sane management of same.

I elected not to modify our metadata format. Although the metadata
files are surely a logical place for this, I'm loath to change the
format so often. Instead, this code reads the type of columns out of
the parquet files for unloaded properties. This is slower but requires
no metadata changes.

We could always add it to the metadata later as an optimization if this
way of doing things is deemed too slow.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>