Python: Enable developers to collect coverage for Python (#424)
Enable developers to collect coverage for Python source code (using Coverage.py). To avoid modifying codebase with sitespecific instructions, we generate usercustomize during the build (rather than having it in the source of the project); the script is generated only if a variable for Coverage.py is set prior to the build.