build: Bound KATANA_NUM_TEST_THREADS
KATANA_NUM_TEST_THREADS is used to to setup test configurations. Since
tests themselves are generally run in parallel (make -j N) and some
tests scale weakly if at all, it is more appropriate for the number of
test threads to be independent of the actual number of processors,
keeping the amount of test work constant. Thus, when we move to larger
machines, tests get faster rather than slower.

Most users of KATANA_NUM_TEST_THREADS already manually bound its value
already to avoid excessive parallelism. This change makes this
consistent for all uses of KATANA_NUM_TEST_THREADS.

The value of 4 was chosen because it permits complicated interleavings,
it is a power of 2 and it is small.