Move bulk of AssignEntityTypeIDsFromProperties to cpp file
The whole function can't be moved because it is templated, but this gets
most of the logic into a cpp file.

An alternative would have been to make this a non-template function and
accept a C-style array instead of a template-defined vector type. Then
the entire function could have been moved to the cpp file.

A variation of this would be to create a class that wraps C-style arrays
and tracks the size and offers begin(), end(), etc.