libtsuba API for optional topology storage
***New Classes***

PartitionTopologyMetadataEntry - This holds the basic metadata for a topology. So things like number of nodes/edge, transpose state, edge/node sort state, and topology file path. These metadata items are needed so users of RDG can quickly ask for a specific form of topology without us having to load each topology into memory. loaded/stored in json.

PartitionTopologyMetadata - Holds all of the PartitionTopologyMetadataEntries. This is managed by the RDGPartHeader

RDGTopology - This is the RDG layers in memory representation of a topology. When we load a graph, each PartitionMetadataEntry gets wrapped in an RDGTopology. It then manages, a FileView, mapping pointers to that FileView, and updating the PartitionMetadataEntry. Binding the FileView to the topology file path only occurs when a user asks for the particular topology. An RDGTopology can also be created from in memory arrays, instead of from a PartitionMetadataEntry, which is how new optional topologies are added.

RDGTopologyManager - This is how RDGCore keeps track of the various RDGTopology objects that are available.

***What the API looks like***
"Load"
1) a user constructs a minimal Shadow RDGTopology. This Shadow only has metadata fields (transpose_state, node/edge sort_state) set.
2) user passes the Shadow RDGTopology to RDG::GetTopology()
3) RDG then asks RDGCore to ask its RDGTopologyManager if it has a RDGTopology that matches the Shadow. The first match is returned.
  - if a match is found, the RDGTopologys topology file is Bind()-ed to the FileView and Mapped
  - if a match is not found, an error is returned up the call stack. The user can choose to ignore this error and move along, or fail if it depends on the topology being available. (like the CSR topology)
4) the user can then use the RDGTopology pointer to access all parts of the topology

"Store"
1) a user ensure that any new RDGTopologies it holds are passed to the RDGTopologyManager by calling RDG::AddTopology(RDGTopology topo). The user then calls RDG::Store like usual.
2) RDG::Store tells RDGTopologyManager to store all RDGTopology objects it holds.
3) a RDGTopology is written to disk if either its path is empty(indicating the rdg has moved, or has not been written) or its file store is invalid
4) RDGTopology writes the available data structures to a new file, and the metadata is passed to its PartitionTopologyMetadataEntry
5) RDGPartHeader writes out the PartitionTopologyMetadata as json

***Changes of note***

- implemented  usage of the RDGTopology API in PropertyGraph GraphTopology to load/store
optional topologies

- rework DropTopology to DropAllTopologies
Since we no longer have a single topology, if we want to drop the
topologies to save memory we must unmap and unbind all of the topologies

- added test for RDGTopology API in libtsuba

- added test for libgalois optional topology loading

JIRA: KAT-574