Fix bug in PropertyGraph::ConstructEntityTypeIDs()
This function was declaring two local variables, node_entity_type_ids_
and edge_entity_type_ids_. These two variables had the exact same names
as two PropertyGraph class members. So the function was operating on the
local objects, not the class members. As a result, all of the work of
the function was discarded after it returned.