tsuba: remove tsuba::Stat
Removed because `tsuba::ListAvailableViews` accomplishes the same thing.
I cleaned up and tweaked ListAvailableViews as well.

The semantics have been tweaked so that a view is returned for the
version 0 of the graph. This "view" (if you can even call it that) is
special in that it's not a valid RDG that can be loaded, but returning
it is a nice way to signal to callers that the graph is empty.

JIRA: KAT-1284

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>