Change LLVM search order in cmake to be the default order. (#515)
Honor the build environment search order for LLVM versions. This may result in using a slightly older version of
llvm-support in cases where multiple are installed, however it enables the environment to select an older versions
intentionally if required. Conda does this and not allowing it to do so causes link errors.