Allow the property cache to evict during RDG initialization (#530)
The cache provides the callback function with the size of the object
being evicted.  The cache is configured with an RDG pointer, which is
vital if the RDG is in the process of being constructed.

Also, eliminate Unload*Property(int i) because properties should
be referred to by name, not position.