tsuba: RDGPartHeader fix partition file regex (#547)
This PR intends to fix a bug found with the deployment. There was an issue where every partition file was being copied over with the same node id. This was because there was an issue with the regex parsing which caused every partition file to be copied over with the same node id of 0, hence the copies were getting butchered. I also wrote a test to catch this edge case.