Apply dynamic_cast consistently
Break the use of dynamic_cast into cases and apply them consistently.

If the cast is assumed to succeed, use either

  auto& ret = dynamic_cast<T&>(item)

or

  auto* ret = dynamic_cast<T*>(item); KATANA_LOG_ASSERT(ret)

In both cases, the program will crash if the assumption does not hold.
This is preferable to memory corruption.

If the cast could fail, use

  auto* ret = dynamic_cast<T*>(item)

And if the cast involves a shared_ptr, use

  auto ret = std::dynamic_pointer_cast<T>(item)

unless the lifetime of item can be guaranteed and the situation requires
extreme performance tuning. dynamic_pointer_cast makes sure that the
lifetime of item and ret are linked and avoids dangling pointers. This
is preferable to memory corruption or a slight performance penalty.