Add (abstract) methods for global degrees to Python Graph
This commit adds global degree methods to the Python graph object. For a
single host, global degree is equivalent to normal degree, but for
distributed execution, global degree allows a user to get the degree of
a node even if the host does not own all of the edges.

The overarching design goal behind this commit is that views that add on
to the base graph (like a bidirectional view) will be generated on
demand whenever a function is called (like global degree) which requires
a view to be built.