fix: always load schema info
One way that we populate our schemas for node and edge properties is by
loading the properties. Another way is by reading the headers of
property files that weren't loaded. We have a bug where if we don't load
any properties then we terminate a function early and skip reading the
headers. Fix that bug.