Fix: tsuba::ListAvailableViews()
The ListAvailableViews() contract is that it will return all views for
the requested version and only views for the requested version. If no
version is requested, it will return all views for the latest version
and only views for the latest version.

The no-requested-view variant of this contract is being violated
because in the core loop over the
potential views there is no check for if the view under consideration is
less than the current most recent known version. So the returned list
contains views from multiple versions.