galois: block_range more evenly
Balance block ranges more evenly. The previous approach of taking the
ceiling can leave the last block much smaller than all others. This
issue has not been significant in practice but it is easy enough to fix
with integers by taking the floor and then adding 1 to the first
remainder blocks.

For example, if we have 13 elements and 4 blocks, the previous approach
of taking the ceiling would result in block sizes of [4, 4, 4, 1]. When
we take the floor and add the remainder, the block sizes are [4, 3, 3,
3]. In the latter approach, the maximum difference in size between any
two blocks is 1, while the previous approach could have a maximum
imbalance of (ceiling(num / blocks) - 1).

Also, move range functions to Range.h.