tsuba: Include property split files in RDGManifest::FilesNames() (#615)
This PR fixes the bug where split property files were not being copied over during a CopyRDGOperation. This was occurring because the RDGManifest only had information for the property file, but excluded the split property files if they existed. This PR tackles this issue by checking to see whether a given property file contains offset mappings or not and then save those file names to be returned later on.

JIRA: https://katanagraph.atlassian.net/browse/KAT-1892