python: Reset busy wait / burnPower value
Without resetting, the leaked value interferes with manual changes of
the number of active threads, which is quite common in tests. Promote
the debug build assert to an always on assertion as inconsistencies in
the number of threads is dangerous.

Along the way, be aggressive about hiding private state of ThreadPool.