PropertyView: bring IsValid() more in line with arrow
The arrow implementation of IsValid checks the null count of the array
when there is no null bitmap. This implies the following model:

An arrow array can omit its null bitmap in two situations: when the
entire array is null and when the entire array is non-null.

It is pretty difficult to actually construct an arrow array
with all null values and no bitmap, but it makes sense to bring the
PODPropertyView implementation of IsValid() in line with the arrow
implementation just to be safe.