OpaqueID: fmt::formatter specialization
The implementation is so complicated because actual opaque ids are
derived classes where the base class is an instance of the OpaqueID
template. The relationship between a used id and the OpaqueID template
may be even more complicated if the id derives from OpaqueIDLinear or
another non-OpaqueID template. Due to some elaborate SFINAE logic, all
opaque ids can now be used with libfmt.

In order to enable this, add IDType as a public using in OpaqueID.