RDG: make prefix and slice compatible with partitioned graphs
RDGPrefix and RDGSlice used to only be used with unpartitioned graphs.
Now they both accept a partition_id parameter so that they can operate
on a single partition within a partitioned graph. This is not a huge
change in their functionality, it just allows them to operate on more
graphs.