query: Add API functions to search for edges satisfying a function
This commit add methods to the GraphTopology.h API.  Each method allows searching through a range of edges until an edge satisfying a boolean function is found.  There is a method for searching through all edges between a source and a destination, all out edges of a node, and all in edges of a node.  

This was motivated by filtering on path expressions.  We need to go backwards from node and edge IDs to edges from edges_range to check edge properties.  This commit aims to simplify this procedure.

jira: KAT-1253
git: #648