Add Undirected PropertyGraph View and use it for Louvain Clustering.  (#666)
This PR provides an Undirected View on top of PropertyGraph, which allows us to run algorithms that expect a symmetric graph input such as Louvain and Connected Components .
Previously, such algorithms required providing a symmetric graph on disk. Now, we can run them with Undirected View on top of a directed graph.
Undirected View == Original Topology + Transpose topology. , combined with an iterator that goes over two ranges (in and out edges of a node).

* add undirected to cc

* changes

* template argument for view

* WIP: Code for undirected view. Doesn't compile

* Rebase on latest master

* Fixed Bugs. Undirected View works

* WIP: rename to UndirectedView

* Bug & formatting fixes. Iterating over two ranges is tricky

* WIP: code cleanup in prep for Undirected View

* WIP: Switch to Undirected View

* WIP: Fixed build. Doesn't terminate

* Minor fixes and debug prints

* Bug fix: all_edges() shouldn't return 2x the edges

* CC uses default view for now

* format fix

* address comments

* fix build

* Fix CI test

* Fix format

Co-authored-by: M. Amber Hassaan <amber_474@yahoo.com>