remove assert from Undirected View (#684)
This PR removes an assertion from UndirectedTopology class that checks if outgoing number of edges is greater than 0. This check is problematic for clustering on very small graphs since we can arise at a situation where the number of edges in a merged/coarsened graph is zero on one of the hosts.