More specific return from divideNodesBinarySearch()
Prior to this commit, this function returned the same empty edge range
for any call with an empty edge range. While all empty edge ranges are
equivalent in a strict sense, a slight modification to this interface
would allow a nice invariant:
If all parameters are kept static except for id, then the end of the
edge range for some id I is equal to the beginning of the edge range for
id I+1

Change the empty edge range behavior of divideNodesBinarySearch() to
maintain this invariant.

JIRA:
GitHub: