support: Remove redundant fmt specialization
fmt picks up on ostream& operator<<(ostream&, T) overloads and there is
typically no need to specialize printing functionality for fmt if
there is already an ostream overload.

This also has the benefit of supporting both fmt::print and std::cout
with the same customization point.

There is an issue with fmt<8 when using fmt::join(items, ...) when Item
only overloads operator<<, which is what prompted the change in
153ab8a80 that this commit reverts. This issue is larger than printing
OpaqueIDs though and specific to fmt::join.