support: Adjust Join API to match fmt::join
Make the API of katana::Join match the API of fmt::join to facilitate
easy refactoring between the two.

Along the way, remove the dependency of katana::Join on fmt. If
katana::Join is a replacement for fmt::join, it would be better if the
former is not dependent on the fmt.