Graph helpers: replace some unsigned with size_t
This doesn't modify large vectors or anything where adding a few bytes
will blow up the memory use.

It is primarily to allow
divideNodesBinarySearch() to accept std::vector<size_t> instead of
std::vector<unsigned> for the scaleFactor parameter. This gives that
parameter some more expressiveness and makes it safer by making it
impossible to pass in a vector of too-large values.