PropStorageInfo: fix state machine transition bug
Is*() functions report state, Was*() functions indicate that state
should transition. Fix the Upsert*() family to use WasModified() (to
indicate that the property should be marked dirty) instead of IsDirty()
(which reports the current state of the property). Also ensure that the
property is marked dirty in both cases (insert and update).