Fix handling of commits with both tags and trivial merges. (#736)
The code simplifies trivial merges to make versions make more
sense when run in github CI (which generates trivial merges).
However, when tags are present it was not checking the original
commit and was only checking the "normalized" commit. This change
makes the system check both the original commit and the
"normalized" commit.

This commit also assumes tags on only enterprise are correct
instead of ignoreing them. It prints a warning, but will still
create the expected version. This should avoid "oops the artifacts
are broken" surprises.