Reset default topology pointer to more constrained topology if available.
This commit has cosmetic and functional changes. The cosmetic change is that EdgeTypeAwareTopology now derives directly from EdgeShuffleTopology. The code in GraphTopology.h has been moved around so that all Topology class definitions precede TopologyWrapper definitions. The functional change is that the default topology has been moved to the topology cache. Moreover, now its pointer may be reseated to point to a more constrained topology, if one is created, to reduce the overall memory consumption of topologies in cache.